feat: integrate fetchData
